+++ src/main/java/com/markokramar/emilfrey/resource/CarsResource.java
@@ -1,7 +1,7 @@
 package com.markokramar.emilfrey.resource;
 
-import com.markokramar.emilfrey.service.CarsService;
 import com.markokramar.emilfrey.model.Car;
+import com.markokramar.emilfrey.service.CarsService;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
@@ -9,8 +9,6 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
-
-import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 
 @RequestScoped
 @Path("cars")
@@ -31,13 +29,29 @@ public class CarsResource {
     @Path("{id}")
     public Response getCar(@PathParam("id") Long id) {
         Car car = carsService.findById(id);
+        if (car == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Car not found for ID: " + id).build();
+        }
+
         return Response.ok(car).build();
     }
 
     @PUT
     @Path("{id}")
     public Response update(@PathParam("id") Long id, Car car) {
+        if (id == null) {
+            return Response.serverError().entity("ID cannot be blank").build();
+        }
+
+        if (car == null) {
+            return Response.serverError().entity("Car param cannot be blank").build();
+        }
+
         Car carToUpdate = carsService.findById(id);
+        if (carToUpdate == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Car not found for ID: " + id).build();
+        }
+
         carToUpdate.setManufacturer(car.getManufacturer());
         carToUpdate.setModel(car.getModel());
         carToUpdate.setCategory(car.getCategory());
@@ -49,13 +63,21 @@ public class CarsResource {
     @POST
     public Response create(Car car) {
         carsService.create(car);
-        return Response.ok().build();
+        return Response.ok(car).build();
     }
 
     @DELETE
     @Path("{id}")
     public Response delete(@PathParam("id") Long id) {
+        if (id == null) {
+            return Response.serverError().entity("ID cannot be blank").build();
+        }
+
         Car carToDelete = carsService.findById(id);
+        if (carToDelete == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Car not found for ID: " + id).build();
+        }
+
         carsService.delete(carToDelete);
         return Response.ok().build();
     }
@@ -63,9 +85,7 @@ public class CarsResource {
     @DELETE
     public Response delete(List<Long> ids) {
         if (ids == null || ids.isEmpty()) {
-            throw new WebApplicationException(
-                    Response.status(BAD_REQUEST).entity("Ids parameter to delete is mandatory").build()
-            );
+            return Response.serverError().entity("IDs need to be provided").build();
         }
 
         carsService.delete(ids);
