+++ src/main/java/com/markokramar/emilfrey/resource/LeadsResource.java
@@ -1,7 +1,7 @@
 package com.markokramar.emilfrey.resource;
 
-import com.markokramar.emilfrey.service.LeadsService;
 import com.markokramar.emilfrey.model.Lead;
+import com.markokramar.emilfrey.service.LeadsService;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
@@ -9,8 +9,6 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
-
-import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 
 @RequestScoped
 @Path("leads")
@@ -31,13 +29,29 @@ public class LeadsResource {
     @Path("{id}")
     public Response getLead(@PathParam("id") Long id) {
         Lead lead = leadsService.findById(id);
+        if (lead == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Lead not found for ID: " + id).build();
+        }
+
         return Response.ok(lead).build();
     }
 
     @PUT
     @Path("{id}")
     public Response update(@PathParam("id") Long id, Lead lead) {
+        if (id == null) {
+            return Response.serverError().entity("ID cannot be blank").build();
+        }
+
+        if (lead == null) {
+            return Response.serverError().entity("Lead param cannot be blank").build();
+        }
+
         Lead leadToUpdate = leadsService.findById(id);
+        if (leadToUpdate == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Lead not found for ID: " + id).build();
+        }
+
         leadToUpdate.setFirstName(lead.getFirstName());
         leadToUpdate.setLastName(lead.getLastName());
         leadToUpdate.setCarsOfInterest(lead.getCarsOfInterest());
@@ -48,13 +62,17 @@ public class LeadsResource {
     @POST
     public Response create(Lead lead) {
         leadsService.create(lead);
-        return Response.ok().build();
+        return Response.ok(lead).build();
     }
 
     @DELETE
     @Path("{id}")
     public Response delete(@PathParam("id") Long id) {
         Lead leadToDelete = leadsService.findById(id);
+        if (leadToDelete == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Lead not found for ID: " + id).build();
+        }
+
         leadsService.delete(leadToDelete);
         return Response.ok().build();
     }
@@ -62,9 +80,7 @@ public class LeadsResource {
     @DELETE
     public Response delete(List<Long> ids) {
         if (ids == null || ids.isEmpty()) {
-            throw new WebApplicationException(
-                    Response.status(BAD_REQUEST).entity("Ids parameter to delete is mandatory").build()
-            );
+            return Response.serverError().entity("IDs need to be provided").build();
         }
 
         leadsService.delete(ids);
